Move ggrc_gdrive_integration's FE code into core module
