+++ src/ggrc/assets/javascripts/components/gdrive/tests/gdrive_picker_launcher_spec.js
