+++ src/ggrc/assets/javascripts/controllers/gapi-modal.js
@@ -3,7 +3,7 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import ModalsController from '../../../../ggrc/assets/javascripts/controllers/modals_controller';
+import ModalsController from './modals_controller';
 
 export default ModalsController({
   pluginName: 'ggrc_controllers_gapi_modal',
