+++ src/ggrc/assets/javascripts/controllers/gapi_controllers.js
