+++ src/ggrc/assets/javascripts/controllers/modals.js
@@ -9,10 +9,7 @@ import './help_controller';
 import './delete_modal_controller';
 import './unmap_modal_controller';
 import './quick_form_controller';
-
-if (IS_GDRIVE_ENABLED) {
-  require('../../../../ggrc_gdrive_integration/assets/javascripts/controllers/gapi-modal');
-}
+import './gapi-modal';
 
 if (IS_WORKFLOWS_ENABLED) {
   require('../../../../ggrc_workflows/assets/javascripts/controllers/approval-workflow-modal');
