+++ src/ggrc/assets/javascripts/entrypoints/modules.js
@@ -5,3 +5,9 @@
 
 // ggrc_basic_permissions
 import '../controllers/contributions';
+
+// ggrc_gdrive_integration
+import '../controllers/gapi_controllers';
+import '../apps/gdrive';
+import '../components/gdrive/gdrive_picker_launcher';
+import '../components/gdrive/gdrive_folder_picker';
