+++ src/ggrc/assets/javascripts/models/index.js
@@ -48,6 +48,6 @@ import './control';
 import './section';
 import './system';
 import './join_models';
-
+import './gdrive_models';
