+++ src/ggrc/assets/mustache/gdrive/confirm_buttons.mustache
