+++ src/ggrc_gdrive_integration/assets/assets.yaml
@@ -2,11 +2,3 @@
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 
 dashboard-js-template-files:
-  - gdrive/auth_button.mustache
-  - gdrive/gapi_errors.mustache
-  - gdrive/gdrive_folder.mustache
-  - gdrive/gdrive_file.mustache
-  - gdrive/confirm_buttons.mustache
-  - roles/gdrive_option_detail.mustache
-  - audits/gdrive_folder_picker.mustache
-  - workflows/gdrive_folder_picker.mustache
