--- src/ggrc_gdrive_integration/assets/javascripts/index.js
@@ -1,10 +1,0 @@
-/*!
-    Copyright (C) 2017 Google Inc.
-    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-*/
-
-import './models/gdrive_models';
-import './controllers/gapi_controllers';
-import './apps/gdrive';
-import './components/gdrive_picker_launcher';
-import './components/gdrive_folder_picker';
